moved serde to file
