+++ crates/primitives/felt/src/lib.rs
@@ -14,6 +14,9 @@ pub extern crate alloc;
 
 mod starkware_types_conversions;
 
+#[cfg(feature = "serde")]
+include!("with_serde.rs");
+
 use alloc::string::{String, ToString};
 
 use cairo_vm::felt::Felt252;
@@ -25,6 +28,9 @@ use sp_core::{H256, U256};
 use starknet_api::hash::StarkFelt;
 use starknet_ff::{FieldElement, FromByteSliceError, FromStrError};
 use thiserror_no_std::Error;
+
+#[cfg(feature = "serde")]
+pub use crate::with_serde::*;
 
 #[derive(Clone, Debug, PartialEq, PartialOrd, Ord, Hash, Eq, Copy)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
@@ -329,11 +335,6 @@ pub enum Felt252WrapperError {
 
 use alloc::borrow::Cow;
 
-#[cfg(feature = "serde")]
-use serde::{Serialize, Serializer};
-#[cfg(feature = "serde")]
-use serde_with::SerializeAs;
-
 impl From<Felt252WrapperError> for Cow<'static, str> {
     fn from(err: Felt252WrapperError) -> Self {
         match err {
@@ -364,19 +365,6 @@ impl From<FromStrError> for Felt252WrapperError {
             FromStrError::InvalidCharacter => Self::InvalidCharacter,
             FromStrError::OutOfRange => Self::OutOfRange,
         }
-    }
-}
-
-#[cfg(feature = "serde")]
-pub struct UfeHex;
-
-#[cfg(feature = "serde")]
-impl SerializeAs<Felt252Wrapper> for UfeHex {
-    fn serialize_as<S>(value: &Felt252Wrapper, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        starknet_core::serde::unsigned_field_element::UfeHex::serialize_as::<S>(&value.0, serializer)
     }
 }
